+++ gluecodium/src/test/resources/smoke/external_types/output/android/com/example/dontsmoke/ExternalMarkedAsSerializable.java
@@ -1,0 +1,12 @@
+/*
+ *
+ */
+package com.example.dontsmoke;
+import android.os.Parcel;
+import android.os.Parcelable;
+final class ExternalMarkedAsSerializable {
+    public int field;
+    ExternalMarkedAsSerializable(final int field) {
+        this.field = field;
+    }
+}